+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.25
+* new function `poincaremap` for iterating over the Poincare map step by step.
 * `mean_return_times` has been improved for continuous systems and now also allows `diffeq...` keyword propagation. The keyword `m` is also deprecated in favor of `dmin`.
 
 # 1.24